pyproject.toml
[tool.poetry]
name = "lunch-helper"
version = "0.1.0"
description = ""
authors = ["jason <jason890088@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
line-bot-sdk = "^3.7.0"
celery = "^5.3.6"
redis = "^5.0.1"
flower = "^2.0.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
selenium = "^4.17.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
